fix : user id를 Long 타입으로 변경
기존에는 int type이였지만 equals비교를 하기 위해 wrapper type으로 변경

+++ Kakao-Backend-Second-DTO/src/main/java/com/example/kakaoshop/user/util/UserDummyData.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class UserDummyData {
-    public User newUser(Integer id, String username) {
+    public User newUser(Long id, String username) {
         PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
         return User.builder()
                 .id(id)
